Use integral numbers in progress of Projects
Everywhere in the system, we expect integral numbers. They are easier to
people understand than any floating point (which doesn't add a lot in a
simple proportion calculation).

+++ db/migrate/20140808185831_create_project_totals.rb
@@ -8,7 +8,7 @@ class CreateProjectTotals < ActiveRecord::Migration
       t.decimal :net_amount, default: 0
       t.decimal :platform_fee, default: 0
       t.decimal :pledged, default: 0
-      t.float :progress, default: 0
+      t.integer :progress, default: 0
       t.integer :total_contributions, default: 0
       t.integer :total_contributions_without_matches, default: 0
       t.decimal :total_payment_service_fee, default: 0
